% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{add_facet_to_object}
\alias{add_facet_to_object}
\title{Adds a new Facet to an object}
\usage{
add_facet_to_object(DirectoryArn, SchemaFacet, ObjectReference,
  ObjectAttributeList = NULL)
}
\arguments{
\item{DirectoryArn}{The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns.}

\item{SchemaFacet}{Identifiers for the facet that you are adding to the object. See SchemaFacet for details.}

\item{ObjectReference}{A reference to the object you are adding the specified facet to.}

\item{ObjectAttributeList}{Attributes on the facet that you are adding to the object.}
}
\description{
Adds a new Facet to an object. An object can have more than one facet applied on it.
}
\examples{

}
