% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{attach_policy}
\alias{attach_policy}
\title{Attaches a policy object to a regular object}
\usage{
attach_policy(
  DirectoryArn = "string",
  PolicyReference = list(
    Selector = "string"
  ),
  ObjectReference = list(
    Selector = "string"
  )
)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) that is associated with the Directory where both objects reside. For more information, see arns.}

\item{PolicyReference}{[required] The reference that is associated with the policy object.}

\item{ObjectReference}{[required] The reference that identifies the object to which the policy will be attached.}
}
\description{
Attaches a policy object to a regular object. An object can have a limited number of attached policies.
}
\examples{

}
