% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{attach_to_index}
\alias{attach_to_index}
\title{Attaches the specified object to the specified index}
\usage{
attach_to_index(DirectoryArn, IndexReference, TargetReference)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) of the directory where the object and index exist.}

\item{IndexReference}{[required] A reference to the index that you are attaching the object to.}

\item{TargetReference}{[required] A reference to the object that you are attaching to the index.}
}
\description{
Attaches the specified object to the specified index.
}
\section{Accepted Parameters}{
\preformatted{attach_to_index(
  DirectoryArn = "string",
  IndexReference = list(
    Selector = "string"
  ),
  TargetReference = list(
    Selector = "string"
  )
)
}
}

