% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{attach_typed_link}
\alias{attach_typed_link}
\title{Attaches a typed link to a specified source and target object}
\usage{
attach_typed_link(DirectoryArn, SourceObjectReference, TargetObjectReference,
  TypedLinkFacet, Attributes)
}
\arguments{
\item{DirectoryArn}{The Amazon Resource Name (ARN) of the directory where you want to attach the typed link.}

\item{SourceObjectReference}{Identifies the source object that the typed link will attach to.}

\item{TargetObjectReference}{Identifies the target object that the typed link will attach to.}

\item{TypedLinkFacet}{Identifies the typed link facet that is associated with the typed link.}

\item{Attributes}{A set of attributes that are associated with the typed link.}
}
\description{
Attaches a typed link to a specified source and target object. For more information, see \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/objectsandlinks.html#typedlink}{Typed link}.
}
\examples{

}
