% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{batch_read}
\alias{batch_read}
\title{Performs all the read operations in a batch}
\usage{
batch_read(DirectoryArn, Operations, ConsistencyLevel = NULL)
}
\arguments{
\item{DirectoryArn}{The Amazon Resource Name (ARN) that is associated with the Directory. For more information, see arns.}

\item{Operations}{A list of operations that are part of the batch.}

\item{ConsistencyLevel}{Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.}
}
\description{
Performs all the read operations in a batch.
}
\examples{

}
