% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{batch_write}
\alias{batch_write}
\title{Performs all the write operations in a batch}
\usage{
batch_write(DirectoryArn, Operations)
}
\arguments{
\item{DirectoryArn}{The Amazon Resource Name (ARN) that is associated with the Directory. For more information, see arns.}

\item{Operations}{A list of operations that are part of the batch.}
}
\description{
Performs all the write operations in a batch. Either all the operations succeed or none.
}
\examples{

}
