% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{create_directory}
\alias{create_directory}
\title{Creates a Directory by copying the published schema into the directory}
\usage{
create_directory(Name, SchemaArn)
}
\arguments{
\item{Name}{The name of the Directory. Should be unique per account, per region.}

\item{SchemaArn}{The Amazon Resource Name (ARN) of the published schema that will be copied into the data Directory. For more information, see arns.}
}
\description{
Creates a Directory by copying the published schema into the directory. A directory cannot be created without a schema.
}
\examples{

}
