% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{create_index}
\alias{create_index}
\title{Creates an index object}
\usage{
create_index(DirectoryArn, OrderedIndexedAttributeList, IsUnique,
  ParentReference = NULL, LinkName = NULL)
}
\arguments{
\item{DirectoryArn}{The ARN of the directory where the index should be created.}

\item{OrderedIndexedAttributeList}{Specifies the attributes that should be indexed on. Currently only a single attribute is supported.}

\item{IsUnique}{Indicates whether the attribute that is being indexed has unique values or not.}

\item{ParentReference}{A reference to the parent object that contains the index object.}

\item{LinkName}{The name of the link between the parent object and the index object.}
}
\description{
Creates an index object. See \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/cd_indexing.html}{Indexing} for more information.
}
\examples{

}
