% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{create_typed_link_facet}
\alias{create_typed_link_facet}
\title{Creates a TypedLinkFacet}
\usage{
create_typed_link_facet(SchemaArn, Facet)
}
\arguments{
\item{SchemaArn}{The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns.}

\item{Facet}{Facet structure that is associated with the typed link facet.}
}
\description{
Creates a TypedLinkFacet. For more information, see \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/objectsandlinks.html#typedlink}{Typed link}.
}
\examples{

}
