% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{delete_directory}
\alias{delete_directory}
\title{Deletes a directory}
\usage{
delete_directory(DirectoryArn)
}
\arguments{
\item{DirectoryArn}{The ARN of the directory to delete.}
}
\description{
Deletes a directory. Only disabled directories can be deleted. A deleted directory cannot be undone. Exercise extreme caution when deleting directories.
}
\examples{

}
