% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{delete_facet}
\alias{delete_facet}
\title{Deletes a given Facet}
\usage{
delete_facet(SchemaArn, Name)
}
\arguments{
\item{SchemaArn}{The Amazon Resource Name (ARN) that is associated with the Facet. For more information, see arns.}

\item{Name}{The name of the facet to delete.}
}
\description{
Deletes a given Facet. All attributes and Rules that are associated with the facet will be deleted. Only development schema facets are allowed deletion.
}
\examples{

}
