% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{delete_object}
\alias{delete_object}
\title{Deletes an object and its associated attributes}
\usage{
delete_object(DirectoryArn, ObjectReference)
}
\arguments{
\item{DirectoryArn}{The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns.}

\item{ObjectReference}{A reference that identifies the object.}
}
\description{
Deletes an object and its associated attributes. Only objects with no children and no parents can be deleted.
}
\examples{

}
