% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{delete_schema}
\alias{delete_schema}
\title{Deletes a given schema}
\usage{
delete_schema(SchemaArn)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) of the development schema. For more information, see arns.}
}
\description{
Deletes a given schema. Schemas in a development and published state can only be deleted.
}
\section{Accepted Parameters}{
\preformatted{delete_schema(
  SchemaArn = "string"
)
}
}

