% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{delete_typed_link_facet}
\alias{delete_typed_link_facet}
\title{Deletes a TypedLinkFacet}
\usage{
delete_typed_link_facet(SchemaArn, Name)
}
\arguments{
\item{SchemaArn}{The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns.}

\item{Name}{The unique name of the typed link facet.}
}
\description{
Deletes a TypedLinkFacet. For more information, see \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/objectsandlinks.html#typedlink}{Typed link}.
}
\examples{

}
