% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{detach_from_index}
\alias{detach_from_index}
\title{Detaches the specified object from the specified index}
\usage{
detach_from_index(DirectoryArn, IndexReference, TargetReference)
}
\arguments{
\item{DirectoryArn}{The Amazon Resource Name (ARN) of the directory the index and object exist in.}

\item{IndexReference}{A reference to the index object.}

\item{TargetReference}{A reference to the object being detached from the index.}
}
\description{
Detaches the specified object from the specified index.
}
\examples{

}
