% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{detach_object}
\alias{detach_object}
\title{Detaches a given object from the parent object}
\usage{
detach_object(DirectoryArn, ParentReference, LinkName)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) that is associated with the Directory where objects reside. For more information, see arns.}

\item{ParentReference}{[required] The parent reference from which the object with the specified link name is detached.}

\item{LinkName}{[required] The link name associated with the object that needs to be detached.}
}
\description{
Detaches a given object from the parent object. The object that is to be detached from the parent is specified by the link name.
}
\section{Accepted Parameters}{
\preformatted{detach_object(
  DirectoryArn = "string",
  ParentReference = list(
    Selector = "string"
  ),
  LinkName = "string"
)
}
}

