% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{enable_directory}
\alias{enable_directory}
\title{Enables the specified directory}
\usage{
enable_directory(
  DirectoryArn = "string"
)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory to enable.}
}
\description{
Enables the specified directory. Only disabled directories can be enabled. Once enabled, the directory can then be read and written to.
}
\examples{

}
