% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{get_directory}
\alias{get_directory}
\title{Retrieves metadata about a directory}
\usage{
get_directory(DirectoryArn)
}
\arguments{
\item{DirectoryArn}{The ARN of the directory.}
}
\description{
Retrieves metadata about a directory.
}
\examples{

}
