% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{get_link_attributes}
\alias{get_link_attributes}
\title{Retrieves attributes that are associated with a typed link}
\usage{
get_link_attributes(DirectoryArn, TypedLinkSpecifier, AttributeNames,
  ConsistencyLevel = NULL)
}
\arguments{
\item{DirectoryArn}{The Amazon Resource Name (ARN) that is associated with the Directory where the typed link resides. For more information, see arns or \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/objectsandlinks.html#typedlink}{Typed link}.}

\item{TypedLinkSpecifier}{Allows a typed link specifier to be accepted as input.}

\item{AttributeNames}{A list of attribute names whose values will be retrieved.}

\item{ConsistencyLevel}{The consistency level at which to retrieve the attributes on a typed link.}
}
\description{
Retrieves attributes that are associated with a typed link.
}
\examples{

}
