% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{get_object_information}
\alias{get_object_information}
\title{Retrieves metadata about an object}
\usage{
get_object_information(DirectoryArn, ObjectReference, ConsistencyLevel = NULL)
}
\arguments{
\item{DirectoryArn}{The ARN of the directory being retrieved.}

\item{ObjectReference}{A reference to the object.}

\item{ConsistencyLevel}{The consistency level at which to retrieve the object information.}
}
\description{
Retrieves metadata about an object.
}
\examples{

}
