% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{get_schema_as_json}
\alias{get_schema_as_json}
\title{Retrieves a JSON representation of the schema}
\usage{
get_schema_as_json(SchemaArn)
}
\arguments{
\item{SchemaArn}{The ARN of the schema to retrieve.}
}
\description{
Retrieves a JSON representation of the schema. See \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/cd_schemas.html#jsonformat}{JSON Schema Format} for more information.
}
\examples{

}
