% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{get_typed_link_facet_information}
\alias{get_typed_link_facet_information}
\title{Returns the identity attribute order for a specific TypedLinkFacet}
\usage{
get_typed_link_facet_information(
  SchemaArn = "string",
  Name = "string"
)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns.}

\item{Name}{[required] The unique name of the typed link facet.}
}
\description{
Returns the identity attribute order for a specific TypedLinkFacet. For more information, see \href{https://docs.aws.amazon.com/clouddirectory/latest/developerguide/directory_objects_links.html#directory_objects_links_typedlink}{Typed Links}.
}
\examples{

}
