% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{list_development_schema_arns}
\alias{list_development_schema_arns}
\title{Retrieves each Amazon Resource Name (ARN) of schemas in the development state}
\usage{
list_development_schema_arns(
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\description{
Retrieves each Amazon Resource Name (ARN) of schemas in the development state.
}
\examples{

}
