% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{list_directories}
\alias{list_directories}
\title{Lists directories created within an account}
\usage{
list_directories(
  NextToken = "string",
  MaxResults = 123,
  state = "ENABLED"|"DISABLED"|"DELETED"
)
}
\arguments{
\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}

\item{state}{The state of the directories in the list. Can be either Enabled, Disabled, or Deleted.}
}
\description{
Lists directories created within an account.
}
\examples{

}
