% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{list_facet_attributes}
\alias{list_facet_attributes}
\title{Retrieves attributes attached to the facet}
\usage{
list_facet_attributes(SchemaArn, Name, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{SchemaArn}{The ARN of the schema where the facet resides.}

\item{Name}{The name of the facet whose attributes will be retrieved.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\description{
Retrieves attributes attached to the facet.
}
\examples{

}
