% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{list_facet_names}
\alias{list_facet_names}
\title{Retrieves the names of facets that exist in a schema}
\usage{
list_facet_names(SchemaArn, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{SchemaArn}{The Amazon Resource Name (ARN) to retrieve facet names from.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\description{
Retrieves the names of facets that exist in a schema.
}
\examples{

}
