% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{list_index}
\alias{list_index}
\title{Lists objects attached to the specified index}
\usage{
list_index(DirectoryArn, IndexReference, RangesOnIndexedValues = NULL,
  MaxResults = NULL, NextToken = NULL, ConsistencyLevel = NULL)
}
\arguments{
\item{DirectoryArn}{The ARN of the directory that the index exists in.}

\item{IndexReference}{The reference to the index to list.}

\item{RangesOnIndexedValues}{Specifies the ranges of indexed values that you want to query.}

\item{MaxResults}{The maximum number of objects in a single page to retrieve from the index during a request. For more information, see \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/limits.html#limits_cd}{AWS Directory Service Limits}.}

\item{NextToken}{The pagination token.}

\item{ConsistencyLevel}{The consistency level to execute the request at.}
}
\description{
Lists objects attached to the specified index.
}
\examples{

}
