% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{list_managed_schema_arns}
\alias{list_managed_schema_arns}
\title{Lists the major version families of each managed schema}
\usage{
list_managed_schema_arns(
  SchemaArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{SchemaArn}{The response for ListManagedSchemaArns. When this parameter is used, all minor version ARNs for a major version are listed.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\description{
Lists the major version families of each managed schema. If a major version ARN is provided as SchemaArn, the minor version revisions in that family are listed instead.
}
\examples{

}
