% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{list_object_parents}
\alias{list_object_parents}
\title{Lists parent objects that are associated with a given object in pagination fashion}
\usage{
list_object_parents(DirectoryArn, ObjectReference, NextToken = NULL,
  MaxResults = NULL, ConsistencyLevel = NULL)
}
\arguments{
\item{DirectoryArn}{The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns.}

\item{ObjectReference}{The reference that identifies the object for which parent objects are being listed.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of items to be retrieved in a single call. This is an approximate number.}

\item{ConsistencyLevel}{Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.}
}
\description{
Lists parent objects that are associated with a given object in pagination fashion.
}
\examples{

}
