% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{list_policy_attachments}
\alias{list_policy_attachments}
\title{Returns all of the \code{ObjectIdentifiers} to which a given policy is attached}
\usage{
list_policy_attachments(DirectoryArn, PolicyReference, NextToken = NULL,
  MaxResults = NULL, ConsistencyLevel = NULL)
}
\arguments{
\item{DirectoryArn}{The Amazon Resource Name (ARN) that is associated with the Directory where objects reside. For more information, see arns.}

\item{PolicyReference}{The reference that identifies the policy object.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of items to be retrieved in a single call. This is an approximate number.}

\item{ConsistencyLevel}{Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.}
}
\description{
Returns all of the \code{ObjectIdentifiers} to which a given policy is attached.
}
\examples{

}
