% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{list_typed_link_facet_names}
\alias{list_typed_link_facet_names}
\title{Returns a paginated list of \code{TypedLink} facet names for a particular schema}
\usage{
list_typed_link_facet_names(SchemaArn, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{SchemaArn}{The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\description{
Returns a paginated list of \code{TypedLink} facet names for a particular schema. For more information, see \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/objectsandlinks.html#typedlink}{Typed link}.
}
\examples{

}
