% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_package.R
\docType{package}
\name{paws.clouddirectory-package}
\alias{paws.clouddirectory}
\alias{paws.clouddirectory-package}
\title{paws.clouddirectory: Amazon CloudDirectory}
\description{
Amazon Cloud Directory is a component of the AWS
Directory Service that simplifies the development and management of
cloud-scale web, mobile, and IoT applications. This guide describes
the Cloud Directory operations that you can call programmatically and
includes detailed information on data types and errors. For
information about AWS Directory Services features, see [AWS Directory
Service](https://aws.amazon.com/directoryservice/) and the [AWS
Directory Service Administration
Guide](http://docs.aws.amazon.com/directoryservice/latest/admin-guide/what_is.html).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
