% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{publish_schema}
\alias{publish_schema}
\title{Publishes a development schema with a major version and a recommended minor version}
\usage{
publish_schema(DevelopmentSchemaArn, Version, MinorVersion = NULL,
  Name = NULL)
}
\arguments{
\item{DevelopmentSchemaArn}{The Amazon Resource Name (ARN) that is associated with the development schema. For more information, see arns.}

\item{Version}{The major version under which the schema will be published. Schemas have both a major and minor version associated with them.}

\item{MinorVersion}{The minor version under which the schema will be published. This parameter is recommended. Schemas have both a major and minor version associated with them.}

\item{Name}{The new name under which the schema will be published. If this is not provided, the development schema is considered.}
}
\description{
Publishes a development schema with a major version and a recommended minor version.
}
\examples{

}
