% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{remove_facet_from_object}
\alias{remove_facet_from_object}
\title{Removes the specified facet from the specified object}
\usage{
remove_facet_from_object(DirectoryArn, SchemaFacet, ObjectReference)
}
\arguments{
\item{DirectoryArn}{The ARN of the directory in which the object resides.}

\item{SchemaFacet}{The facet to remove. See SchemaFacet for details.}

\item{ObjectReference}{A reference to the object to remove the facet from.}
}
\description{
Removes the specified facet from the specified object.
}
\examples{

}
