% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{An API operation for adding tags to a resource}
\usage{
tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource. Tagging is only supported for directories.}

\item{Tags}{[required] A list of tag key-value pairs.}
}
\description{
An API operation for adding tags to a resource.
}
\section{Accepted Parameters}{
\preformatted{tag_resource(
  ResourceArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

