% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{An API operation for removing tags from a resource}
\usage{
untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{The Amazon Resource Name (ARN) of the resource. Tagging is only supported for directories.}

\item{TagKeys}{Keys of the tag that need to be removed from the resource.}
}
\description{
An API operation for removing tags from a resource.
}
\examples{

}
