% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{update_facet}
\alias{update_facet}
\title{Does the following:}
\usage{
update_facet(SchemaArn, Name, AttributeUpdates = NULL, ObjectType = NULL)
}
\arguments{
\item{SchemaArn}{The Amazon Resource Name (ARN) that is associated with the Facet. For more information, see arns.}

\item{Name}{The name of the facet.}

\item{AttributeUpdates}{List of attributes that need to be updated in a given schema Facet. Each attribute is followed by \code{AttributeAction}, which specifies the type of update operation to perform.}

\item{ObjectType}{The object type that is associated with the facet. See CreateFacetRequest\$ObjectType for more details.}
}
\description{
Does the following:
}
\details{
\enumerate{
\item Adds new \code{Attributes}, \code{Rules}, or \code{ObjectTypes}.
\item Updates existing \code{Attributes}, \code{Rules}, or \code{ObjectTypes}.
\item Deletes existing \code{Attributes}, \code{Rules}, or \code{ObjectTypes}.
}
}
\examples{

}
