% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{update_link_attributes}
\alias{update_link_attributes}
\title{Updates a given typed link's attributes}
\usage{
update_link_attributes(DirectoryArn, TypedLinkSpecifier, AttributeUpdates)
}
\arguments{
\item{DirectoryArn}{The Amazon Resource Name (ARN) that is associated with the Directory where the updated typed link resides. For more information, see arns or \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/objectsandlinks.html#typedlink}{Typed link}.}

\item{TypedLinkSpecifier}{Allows a typed link specifier to be accepted as input.}

\item{AttributeUpdates}{The attributes update structure.}
}
\description{
Updates a given typed link's attributes. Attributes to be updated must not contribute to the typed link's identity, as defined by its \code{IdentityAttributeOrder}.
}
\examples{

}
