% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{update_schema}
\alias{update_schema}
\title{Updates the schema name with a new name}
\usage{
update_schema(
  SchemaArn = "string",
  Name = "string"
)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) of the development schema. For more information, see arns.}

\item{Name}{[required] The name of the schema.}
}
\description{
Updates the schema name with a new name. Only development schema names can be updated.
}
\examples{

}
