% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{update_typed_link_facet}
\alias{update_typed_link_facet}
\title{Updates a TypedLinkFacet}
\usage{
update_typed_link_facet(SchemaArn, Name, AttributeUpdates,
  IdentityAttributeOrder)
}
\arguments{
\item{SchemaArn}{The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns.}

\item{Name}{The unique name of the typed link facet.}

\item{AttributeUpdates}{Attributes update structure.}

\item{IdentityAttributeOrder}{The order of identity attributes for the facet, from most significant to least significant. The ability to filter typed links considers the order that the attributes are defined on the typed link facet. When providing ranges to a typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range. Filters are interpreted in the order of the attributes on the typed link facet, not the order in which they are supplied to any API calls. For more information about identity attributes, see \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/objectsandlinks.html#typedlink}{Typed link}.}
}
\description{
Updates a TypedLinkFacet. For more information, see \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/objectsandlinks.html#typedlink}{Typed link}.
}
\examples{

}
