% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{apply_schema}
\alias{apply_schema}
\title{Copies the input published schema, at the specified version, into the Directory with the same name and version as that of the published schema}
\usage{
apply_schema(PublishedSchemaArn, DirectoryArn)
}
\arguments{
\item{PublishedSchemaArn}{Published schema Amazon Resource Name (ARN) that needs to be copied. For more information, see arns.}

\item{DirectoryArn}{The Amazon Resource Name (ARN) that is associated with the Directory into which the schema is copied. For more information, see arns.}
}
\description{
Copies the input published schema, at the specified version, into the Directory with the same name and version as that of the published schema.
}
\examples{

}
