% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{attach_object}
\alias{attach_object}
\title{Attaches an existing object to another object}
\usage{
attach_object(
  DirectoryArn = "string",
  ParentReference = list(
    Selector = "string"
  ),
  ChildReference = list(
    Selector = "string"
  ),
  LinkName = "string"
)
}
\arguments{
\item{DirectoryArn}{[required] Amazon Resource Name (ARN) that is associated with the Directory where both objects reside. For more information, see arns.}

\item{ParentReference}{[required] The parent object reference.}

\item{ChildReference}{[required] The child object reference to be attached to the object.}

\item{LinkName}{[required] The link name with which the child object is attached to the parent.}
}
\description{
Attaches an existing object to another object. An object can be accessed in two ways:
}
\details{
\enumerate{
\item Using the path
\item Using \code{ObjectIdentifier}
}
}
\examples{

}
