% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{detach_policy}
\alias{detach_policy}
\title{Detaches a policy from an object}
\usage{
detach_policy(
  DirectoryArn = "string",
  PolicyReference = list(
    Selector = "string"
  ),
  ObjectReference = list(
    Selector = "string"
  )
)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) that is associated with the Directory where both objects reside. For more information, see arns.}

\item{PolicyReference}{[required] Reference that identifies the policy object.}

\item{ObjectReference}{[required] Reference that identifies the object whose policy object will be detached.}
}
\description{
Detaches a policy from an object.
}
\examples{

}
