% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{detach_typed_link}
\alias{detach_typed_link}
\title{Detaches a typed link from a specified source and target object}
\usage{
detach_typed_link(DirectoryArn, TypedLinkSpecifier)
}
\arguments{
\item{DirectoryArn}{The Amazon Resource Name (ARN) of the directory where you want to detach the typed link.}

\item{TypedLinkSpecifier}{Used to accept a typed link specifier as input.}
}
\description{
Detaches a typed link from a specified source and target object. For more information, see \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/objectsandlinks.html#typedlink}{Typed link}.
}
\examples{

}
