% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{disable_directory}
\alias{disable_directory}
\title{Disables the specified directory}
\usage{
disable_directory(DirectoryArn)
}
\arguments{
\item{DirectoryArn}{The ARN of the directory to disable.}
}
\description{
Disables the specified directory. Disabled directories cannot be read or written to. Only enabled directories can be disabled. Disabled directories may be reenabled.
}
\examples{

}
