% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{get_applied_schema_version}
\alias{get_applied_schema_version}
\title{Returns current applied schema version ARN, including the minor version in use}
\usage{
get_applied_schema_version(SchemaArn)
}
\arguments{
\item{SchemaArn}{The ARN of the applied schema.}
}
\description{
Returns current applied schema version ARN, including the minor version in use.
}
\examples{

}
