% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{get_object_attributes}
\alias{get_object_attributes}
\title{Retrieves attributes within a facet that are associated with an object}
\usage{
get_object_attributes(DirectoryArn, ObjectReference, SchemaFacet,
  AttributeNames, ConsistencyLevel = NULL)
}
\arguments{
\item{DirectoryArn}{The Amazon Resource Name (ARN) that is associated with the Directory where the object resides.}

\item{ObjectReference}{Reference that identifies the object whose attributes will be retrieved.}

\item{SchemaFacet}{Identifier for the facet whose attributes will be retrieved. See SchemaFacet for details.}

\item{AttributeNames}{List of attribute names whose values will be retrieved.}

\item{ConsistencyLevel}{The consistency level at which to retrieve the attributes on an object.}
}
\description{
Retrieves attributes within a facet that are associated with an object.
}
\examples{

}
