% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{get_object_information}
\alias{get_object_information}
\title{Retrieves metadata about an object}
\usage{
get_object_information(DirectoryArn, ObjectReference,
  ConsistencyLevel = NULL)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory being retrieved.}

\item{ObjectReference}{[required] A reference to the object.}

\item{ConsistencyLevel}{The consistency level at which to retrieve the object information.}
}
\description{
Retrieves metadata about an object.
}
\section{Accepted Parameters}{
\preformatted{get_object_information(
  DirectoryArn = "string",
  ObjectReference = list(
    Selector = "string"
  ),
  ConsistencyLevel = "SERIALIZABLE"|"EVENTUAL"
)
}
}

