% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{list_attached_indices}
\alias{list_attached_indices}
\title{Lists indices attached to the specified object}
\usage{
list_attached_indices(DirectoryArn, TargetReference, NextToken = NULL,
  MaxResults = NULL, ConsistencyLevel = NULL)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory.}

\item{TargetReference}{[required] A reference to the object that has indices attached.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}

\item{ConsistencyLevel}{The consistency level to use for this operation.}
}
\description{
Lists indices attached to the specified object.
}
\section{Accepted Parameters}{
\preformatted{list_attached_indices(
  DirectoryArn = "string",
  TargetReference = list(
    Selector = "string"
  ),
  NextToken = "string",
  MaxResults = 123,
  ConsistencyLevel = "SERIALIZABLE"|"EVENTUAL"
)
}
}

