% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.clouddirectory_operations.R
\name{put_schema_from_json}
\alias{put_schema_from_json}
\title{Allows a schema to be updated using JSON upload}
\usage{
put_schema_from_json(SchemaArn, Document)
}
\arguments{
\item{SchemaArn}{The ARN of the schema to update.}

\item{Document}{The replacement JSON schema.}
}
\description{
Allows a schema to be updated using JSON upload. Only available for development schemas. See \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/cd_schemas.html#jsonformat}{JSON Schema Format} for more information.
}
\examples{

}
